* ♻️ [refactor] 파일 구조 리팩토링, 안쓰는 주석 부분 삭제
