--- TripLog/TripLog/Feat-Stone/Model/TestTodayDummyData.swift
@@ -1,16 +1,0 @@
-//
-//  Expense.swift
-//  TripLog
-//
-//  Created by 김석준 on 1/20/25.
-//
-
-import Foundation
-
-struct TestTodayExpense {
-    let date: String        // 날짜
-    let title: String       // 제목
-    let category: String    // 카테고리
-    let amount: String     // 지출 금액
-    let exchangeRate: String // 환율 금액
-}
